Extract name/age to a concern
Both Visitors and Prisoners have first name, last name, and date of
birth fields with validations, and compute full name, last initial, and
age from these. Extracting this to a module (called Person) eliminates
duplication and subtle differences in behaviour.

This eliminates the last remaining incorrect age calculation from the
application. I think.
